star_bert/pyproject.toml
[project]
name = "starmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==0.28.0",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.11.7",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "attrs==24.2.0",
    "blis==1.0.1",
    "catalogue==2.0.10",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "cloudpathlib==0.20.0",
    "confection==0.1.5",
    "cymem==2.0.8",
    "datasets==3.1.0",
    "dill==0.3.8",
    "evaluate==0.4.3",
    "filelock==3.16.1",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "huggingface-hub==0.26.2",
    "idna==3.10",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib>=3.10.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "murmurhash==1.0.10",
    "networkx==3.4.2",
    "numpy==2.0.2",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "packaging==24.2",
    "pandas==2.2.3",
    "preshed==3.0.9",
    "propcache==0.2.0",
    "protobuf==5.28.3",
    "psutil==6.1.0",
    "pyarrow==18.0.0",
    "pydantic==2.10.0",
    "pydantic-core==2.27.0",
    "pygments==2.18.0",
    "pyqt5>=5.15.11",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "seaborn>=0.13.2",
    "setuptools==75.6.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smart-open==7.0.5",
    "spacy==3.8.2",
    "spacy-alignments==0.9.1",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "spacy-transformers==1.3.5",
    "srsly==2.4.8",
    "sympy==1.13.1",
    "tensorboardx==2.6.2.2",
    "thinc==8.3.2",
    "threadpoolctl==3.5.0",
    "tokenizers==0.15.2",
    "torch==2.5.1",
    "tqdm==4.67.0",
    "transformers==4.36.2",
    "triton==3.1.0",
    "typer==0.13.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.16.0",
    "xxhash==3.5.0",
    "yarl==1.18.0",
]
